[SYCL] Split sycl::half into an internal impl header
Move the sycl::half implementation into detail/half_type_impl.hpp and keep
sycl/half_type.hpp as the public wrapper for stream operators and std::hash.

This lets internal headers, including builtins, depend on the half
implementation without pulling in the full public half header that relies on
std::optional and iostream.

This reduces compile time by another 30ms on builtins.
