+++ sycl/test/include_deps/sycl_khr_includes_group_algorithms.hpp.cpp
@@ -37,6 +37,7 @@
 // CHECK-NEXT: device_event.hpp
 // CHECK-NEXT: pointers.hpp
 // CHECK-NEXT: half_type.hpp
+// CHECK-NEXT: detail/half_type_impl.hpp
 // CHECK-NEXT: bit_cast.hpp
 // CHECK-NEXT: aspects.hpp
 // CHECK-NEXT: info/aspects.def
