+++ sycl/test/include_deps/sycl_khr_includes_half.hpp.cpp
@@ -9,8 +9,8 @@
 // CHECK-NEXT: detail/defines_elementary.hpp
 // CHECK-NEXT: feature_test.hpp
 // CHECK-NEXT: half_type.hpp
+// CHECK-NEXT: detail/half_type_impl.hpp
 // CHECK-NEXT: bit_cast.hpp
-// CHECK-NEXT: detail/export.hpp
 // CHECK-NEXT: detail/fwd/half.hpp
 // CHECK-NEXT: aspects.hpp
 // CHECK-NEXT: detail/defines.hpp
