+++ sycl/test/include_deps/sycl_khr_includes_kernel_bundle.hpp.cpp
@@ -35,6 +35,7 @@
 // CHECK-NEXT: aliases.hpp
 // CHECK-NEXT: bit_cast.hpp
 // CHECK-NEXT: half_type.hpp
+// CHECK-NEXT: detail/half_type_impl.hpp
 // CHECK-NEXT: detail/fwd/half.hpp
 // CHECK-NEXT: ext/oneapi/matrix/matrix-unified-utils.hpp
 // CHECK-NEXT: __spirv/spirv_types.hpp
