+++ sycl/test/include_deps/sycl_khr_includes_usm.hpp.cpp
@@ -13,6 +13,13 @@
 // CHECK-NEXT: detail/builtins/builtins.hpp
 // CHECK-NEXT: detail/fwd/multi_ptr.hpp
 // CHECK-NEXT: access/access.hpp
+// CHECK-NEXT: detail/half_type_impl.hpp
+// CHECK-NEXT: bit_cast.hpp
+// CHECK-NEXT: detail/fwd/half.hpp
+// CHECK-NEXT: aspects.hpp
+// CHECK-NEXT: detail/defines.hpp
+// CHECK-NEXT: info/aspects.def
+// CHECK-NEXT: info/aspects_deprecated.def
 // CHECK-NEXT: detail/loop.hpp
 // CHECK-NEXT: detail/type_traits.hpp
 // CHECK-NEXT: detail/type_traits/bool_traits.hpp
@@ -21,22 +28,15 @@
 // CHECK-NEXT: detail/named_swizzles_mixin.hpp
 // CHECK-NEXT: detail/vector_traits.hpp
 // CHECK-NEXT: aliases.hpp
-// CHECK-NEXT: bit_cast.hpp
 // CHECK-NEXT: detail/common.hpp
 // CHECK-NEXT: detail/assert.hpp
 // CHECK-NEXT: __spirv/spirv_vars.hpp
 // CHECK-NEXT: detail/export.hpp
 // CHECK-NEXT: detail/nd_loop.hpp
 // CHECK-NEXT: detail/fwd/accessor.hpp
-// CHECK-NEXT: detail/defines.hpp
-// CHECK-NEXT: detail/fwd/half.hpp
 // CHECK-NEXT: detail/generic_type_traits.hpp
 // CHECK-NEXT: detail/type_traits/integer_traits.hpp
 // CHECK-NEXT: detail/memcpy.hpp
-// CHECK-NEXT: half_type.hpp
-// CHECK-NEXT: aspects.hpp
-// CHECK-NEXT: info/aspects.def
-// CHECK-NEXT: info/aspects_deprecated.def
 // CHECK-NEXT: marray.hpp
 // CHECK-NEXT: detail/builtins/common_functions.inc
 // CHECK-NEXT: detail/builtins/helper_macros.hpp
@@ -60,6 +60,7 @@
 // CHECK-NEXT: ext/oneapi/experimental/forward_progress.hpp
 // CHECK-NEXT: ext/oneapi/matrix/query-types.hpp
 // CHECK-NEXT: ext/oneapi/bfloat16.hpp
+// CHECK-NEXT: half_type.hpp
 // CHECK-NEXT: ext/oneapi/matrix/matrix-unified-utils.hpp
 // CHECK-NEXT: __spirv/spirv_types.hpp
 // CHECK-NEXT: info/platform_traits.def
